use intptr_t instead of int in some casts to fix 64bit compilation issues
git-svn-id: https://supercollider.svn.sourceforge.net/svnroot/supercollider/trunk@5148 a380766d-ff14-0410-b294-a243070f3f08

+++ source/server/SC_ComPort.cpp
@@ -324,7 +324,7 @@ void DumpReplyAddress(ReplyAddress *inReplyAddress)
 	scprintf("mSockAddr.sin_family %d\n", inReplyAddress->mSockAddr.sin_family);
 	scprintf("mSockAddr.sin_port %d\n", inReplyAddress->mSockAddr.sin_port);
 	scprintf("mSockAddr.sin_addr.s_addr %d\n", inReplyAddress->mSockAddr.sin_addr.s_addr);
-	scprintf("mReplyFunc %08X\n", (uint32)inReplyAddress->mReplyFunc);
+	scprintf("mReplyFunc %08X\n", (intptr_t)inReplyAddress->mReplyFunc);
 }
 
 /*
